requirements.txt
openai
twilio
python-dotenv
flask
pyngrok